Reorganize loaders/deduplicate
