+++ src/app/embeds/(modules)/pomodoro/metadata.ts
@@ -1,9 +1,9 @@
-import type { IModuleMetadata } from '~/lib/embeds/modules';
+import type { IModuleMetadata } from '~/lib/modules/';
 
 const metadata: IModuleMetadata = {
   name: 'Pomodoro',
   description: 'Pomodoro timer which can be embedded in any application',
-  tags: [],
+  tags: ['productivity'],
 };
 
 export default metadata;
