+++ src/lib/modules/apps/registerModules.ts
