+++ src/lib/modules/embeds/registerModules.ts
