+++ src/lib/modules/index.ts
@@ -1,0 +1,16 @@
+export type ITag =
+  | 'document'
+  | 'productivity'
+  | 'converter'
+  | 'fun'
+  | 'miscellaneous';
+export interface IModuleMetadata {
+  name: string;
+  description: string;
+  tags: ITag[];
+}
+
+export interface IModule {
+  route_name: string;
+  metadata: IModuleMetadata;
+}
